Add configureReentrantLocalization to @lit/localize
